backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.6,<4.7
psycopg[binary]>=3.2,<3.3
celery>=5.4,<5.5
redis>=5.2,<5.3
python-decouple>=3.8,<3.9
gunicorn>=23.0,<24.0
